ETM: Removed FirstList hack.
This was never working with a mimetype filter because the toplevel collections
would never match the filter and the initial fetch was thus always empty.
It used to be fixed automatically because the second fetch used to be
a recursive fetch from root anyways.

Simply listing all collections right away should be fast enough, and removing
the hack allowed me to fold a bunch of codepaths into one.