ETM: Split fetchJobDone + correctly emit dataChanged signals.
Split fetchJobDone into separate versions for Item/Collection fetch jobs.
Having a combined slot was confusing since the codepaths were separate anyways.
Additionally dataChanged was emitted in the wrong slot, as it is required
to notify about the changed fetch state, so it has to be emitted after the
itemFetchJob.