ETM: Don't ignore itemChanged of items in virtual collections
It can happen that storage collection of an item is not in the ETM, but
only virtual collections containing items from that collection (see Gmail
resource for example). In that case ETM will only receive itemChanged
notifications for the virtual collections, but not for the storage one,
so the item won't be updated, because ETM does not update the original
item when the notification is detined for a virtual collection.

This change makes ETM not to ignore itemChange notifications when the
collection is virtual, but perserves the orignal parent collection.

BUG: 337665
Approved-By: Christian Mollekopf