Use QMap instead of QHash to store data in DataDescription schema type
When iterating over QHash, the order of entries is not guarenteed, so the
DbInitializerTest (which relies on columns being always returned in the same
order) started failing in Qt 5, probably because of implementation change in
QHash.

This change replaces QHash with QMap, which guarantees ordering when iterating
over it. There is no reason to use QHash in the first place, since we never do
lookup, only iteration (and the data set is very very small)

Also adapts DbInitializerTest test data to the new behavior which ensures that
values are INSERTed into DB in alphabetical order of column names.

This change makes DbInitializerTest pass in Qt 5 build.