Disable QVariant::Bool support in Qt 5 builds of QSQLITE3 driver
Apparently there has been some behavior change in QtSql in Qt 5 that breaks
our use of boolean in SQLite. Problem with SQLite is that it allows you to
store 1 or 'true' into BOOL column as valid values, but when querying them,
WHERE val = 1 will return different results than WHERE val = 'true'.

In Qt 4 the value is always correctly converted to 'true' and 'false' in both
INSERTS and SELECTs, but in Qt 5 it is not.

This needs a global fix in Akonadi to make sure that we always use the string
values everywhere. Until then, I disabled the Boolean support in the SQLite driver
in Qt 4 to make sure at least Qt 4 build of Akonadi works correctly with QSQLITE3.

The problem can be reproduced with the native QSQLITE driver in Qt 5 too.