Make sure we always use 1 and 0 for boolean values in SQLite
In SQLite it's possible to use either 1/0 or true/false to represent
boolean values. However mixing them together in one table does not work,
as WHERE column = 1 and WHERE column = 'true' returns different results.

This worked in Qt 4 for some reason, but does no longer work in Qt 5, so
we need to make sure that we always use the numberical values to represent
boolean values in SQLite everywhere in Akonadi.

With this change, all unit-tests in Qt 4 and Qt 5 builds finally pass again
and we can use the updated fork of QSQLITE driver.

No other backends are affected.