Don't mass-modify committed changes from ResourceBase
The resource could have changed RIDs of the items, and mass-changing
RIDs is not supported by the server, so the RID change would never be
applied. Instead we run an ItemModifyJob for each item passed to
changesCommitted() and wrap them into single TransactionSequence.

This should fix some bugs with item moves on IMAP.

Reviewed-by: Christian Mollekopf