Simplify the Item payload template magic
This removes the rather crappy workaround for iterating over list of
shared_pointer_traits by iterating from first to last. Instead the last
trait points to the first one, so we can easily iterate over the whole
list until we reach the starting point again. This is safer and nicer.