Avoid recursive collection listing in SearchHelper
The recursive listing generates one SQL query per collection, and since search
is invoked rather often (basically whenever you open an email in KMail), we get
lots of unnecessary queries. This new algorithm does one query to fetch all
folders with matching mime types, and only falls back to query the parent chain
when the requested ancestor is not 0 (root), or when the result collection is
not a direct descendant of the requested ancestor.