ChangeRecorder: Don't get stuck on invalid notifications.
This is a manual partial revert of
a22afe7: Fixed endless loop in changerecorder with unhandled notifications.

The previous patch failed to skip over notifications that fail to be emitted in the translateAndCompress branch of replayNext.
This resulted in stuck resources. We still have to keep the dequeueNotification to
avoid endlessly replaying the same broken notification.