StatisticsProxyModel: get valid rightBottom index in proxyDataChanged()
Using QModelIndex::sibling() would return invalid index when the column number
is higher than column count in the source model (because it would internally
call QIdentityProxyModel::sibling(), which does not know about the new columns
we added in StatisticsProxyModel). Using our own index() implementation to get
the index is the way to go.

This fixed crash when selecting folder in KMail (Qt 5 only), because we were
hitting an assert in KRecursiveFilterProxyModel.