Workaround an extrem performance regression in QIcon::fromTheme()
QIcon::fromTheme() calls (in KDE at least) code from FrameworkIntegration plugin,
which does lots and lots of stat() calls just to find proper size of the icon.
This makes the call extremly slow and KMail's collection view extremly sluggish.

To workaround the problem temporarily, we cache the icons in CollectionModel
(it's only a few icons so the cache has only a very small memory footprint.

CCBUG: 346644