Add the QIcon::fromTheme() workaround to ETM
Workaroud the very slow QIcon::fromTheme() directly in ETM, which gives us the
cache for free for both Collections and Items. The cache is really small (not
more than 10 icons) so we can afford to have it all the way down in ETM.

Also adjusted StatisticsProxyModel to call ETM::data() for DecorationRole
instead of trying to resolve the icon on its own.

CCBUG: 346644