Formalize the protocol via QDataStream serializers
Each command and response for the protocol is now represented by a class
with respective member variables and setters/getters as well as overloaded
streaming operators for QDataStream.

The classes are in the libKF5AkonadiPrivate library, which is shared between
the server and the client libraries. This effectively enforces both sides to
use the code to (de)serialize commands.