Remove ClientCapabilities support
We no longer support backwards-compatibility with older clients, so there's
no need for clients announcing their capabilities, as we can always be sure
that the client supports the same features as the server.

This also drops support for older versions of NotificationMessages - they are
no longer emitted, but the API is still there (since it's not strictly related
to binary protocol change and can be removed later on in master).