Fix SQL error with PGSQL when creating new Relations
We cannot use Relation::insert() to create new Relations as that generates
invalid query "INSERT INTO RelationTable  .... RETURNING id" but there's
no id column in RelationTable. This is limitation of the generated Entity
classes and I don't feel like adding another exception into the XSLD, so we
build the queries manually with QueryBuilder.