Improve payload streaming implementation
To improve performance of clients slightly the payload streaming process is now
more similar to the old one. Server first requests metadata (size, version) of
the payload part, which client provides. Then server decides whether to stream
to file or to database and sends a command to client with relevant information.
Finally the client either writes the payload to the file provided in the server
command, or sends the data directly to server.