Use custom QDataStream implementation for streaming
QDataStream unfortunately does not support waiting for data when it reads
past the end of the stream, instead of throws an error and removes the data
it has read so far. This does not suit as very well, because we don't know
the size of the whole Command in advance and thus can't ensure that enough
data is buffered locally for QDataStream to serialize the whole command
succesfully. The trick we've been using up until now was to serialize the
command into QByteArray, and then send size of the QByteArray first before
sending the array, which allows the receiving side to wait for all data.

This is however slow and uses more memory, so the effectivness or binary
protocol is lost. So instead of that we now have Akonadi::Protocol::DataStream
which implements streaming for some elementary Qt and POD types that we
use in Protocol. Unlike QDataStream it always ensures that it has enough
data buffered in the underlaying stream before trying to deserialize them
(it also does more effective container deserialization).

Most of the (de)serialization code is based on the code from QDataStream
(minus backwards compatibility).