SearchManager: fix deadlock on search update
When searchUpdateImpl() is called it can sometimes finish faster than the caller
gets to the QWaitCondition::wait() call. In such situation the searchUpdateImpl()
have already woken the condition and returned, so when caller reaches the wait()
call, it will wait forever. To workaround this we now use QSemaphore to ensure that
caller does not wait for searchUpdateImpl() if it has already returned.