Increase the forced event processing to every 1000 iterations
ItemFetchJob on folder with 40000 emails takes 7 seconds without the forced event
processing and almost 100 seconds with forced processing every 100 iterations.
Increasing the limit to every 1000 iterations reduced the FetchJob duration to 25
seconds.

The major difference is in when user sees the first emails in KMail. In the first
case it is about 15 seconds, in the second it's about 4 seconds, in the last case
it's around 7 seconds.

On the other hand when we serve all emails at once to ETM, the threading performs
*MUCH* faster than if we feed it emails in smaller chunks. 1000 iterations seems
to be more or less good balance between those two, but there's still room for finer
tuning.