Prepare for removal of boost from kdepimlibs
We still need boost in itempayloadinternals_p.h for shared_pointer_trait
and for topological_sort in typepluginloader.cpp.

I guess we can't remove the topological_sort, but that's internal. To
remove from itempayloadinternals, we need to come up with a clever
way to generate the trait only when boost is available.