ItemRetriever: remove duplicate part names
Originally the code was using QSet, which took care of it for us,
but using QVector is more efficient in this case, so we need to
make sure that the vector of part names we get does not contain
duplicates, otherwise the code in exec() asserts.

The code should be quite effective, since the vector usually
contains only one or two part types.

Discovered by itemfetchtest in kdepimlibs.