Fix Agents and Resources not quitting when asked to and crashing instead
The port from KGlobal::ref() to QEventLoopLocker was not working as expected,
since QEventLoopLocker will only quit the application when quitOnLastWindowClosed
is set to true. Otherwise nothing would happen when the reference counter
dropped to 0. Repeated calls to quit() then caused the mEventLoopLocker to be
deleted multiple times and thus crashing.