Slightly speed up ItemRetriever main query (remove two JOINs)
Don't join MimeTypeTable and ResourceTable, instead use the same trick as with
mimetype and flags in FetchHelper, that is local name cache and T::retrieveById().
This is not much, but on large folders it should slightly reduce the query
execution time. The major bottleneck here is still the sorting, but we cannot
really get rid of it.

See T628 on Phabricator for more details and measurements.