Make AkonadiServer QObject, manage multiple QLocalServers for different purposes
This creates so called CommandServer and NotificationServer, both
QLocalServer listening using named pipes (this is a move away from
a socket file to avoid having multiple files).

Each new connection on CommandServer makes AkonadiServer to instantiate
new Connection object as normally. For NotificationServer each new
connection is handed over to NotificationManager, which will deal with
it.