Drop mysql.conf option removed in MySQL 5.7
innodb_additional_mem_pool_size option has been deprecated in MySQL 5.6
and removed in 5.7 so MySQL will no longer start it the option is present
in mysql.conf.

The option still seems to be supported by MariaDB, but the default value
we had in the mysql.conf was below the minimum allowed value, so the harm
in removing the option for MariaDB should be minimal.

If options supported by MySQL and MariaDB begin to diverge more in the future
we might consider having separate config files for each database.