Monitor: translate batch inter-resource move notifications
When we get an inter-resource move notification with multiple Items
we must always run it through translateAndCompress() to translate it
into a series of Add notifications on destination resource and Remove
notification on the source resource.

Without the transformation any inter-resource move of more than one
Item was still being emitted as Monitor::itemsMoved() possibly
causing the resource to fail the operation. In case of non-batch
resources like Maildir resource it only emitted the first Item via
Monitor::itemMoved().

This fixes a potential data loss as the change was not handled by the
destination resource and caused the Items to be stored only in the
Akonadi database.

BUG: 339181
FIXED-IN: 16.04