Fix assert for translated inter-resource remove of multiple items
The special handling to ensure translation of inter-resource move
did not consider moving of multiple items and the fact that when
it gets translated into Remove on source resource (or client) the
resource might not support batch removal.

Make sure that if split is needed we first perform split, then
translate to Remove/Add.

BUG: 361263
FIXED-IN: 16.04.1