Fix MySQL initialization with NO_ZERO_DATE mode
We don't enable NO_ZERO_DATE by default, but external servers may be
configured to use it and will fail to create the PimItemTable in such
case.

Having a default time on atime makes no difference since we already
explicitly set atime to QDateTime::currentDateTime() in the code
when appending a new PimItem.

BUG: 362580
FIXED-IN: 16.04.2