Fix support for newer PostgreSQL versions, make the lookup generic
On most distros PostgreSQL is installed in /usr/lib/postgresql/X.Y/
(where X.Y is the version of the server). Constantly manually updating
the internal list of all possible versions does not really scale, so
instead the code now lists all folders in /usr/lib/postgresql/, sorts
it in descending order and then looks for the first folder that contains
pg_ctl.