Jobs: make sure all subjobs are finished before emitting result()
This affects mostly CollectionSync, which would emit result() signal
while a CommitTransaction subjob was still running. This caused the
response to the CommitTransaction to be delivered to subsequent job
started after the CollectionSync and was causing all kinds of havoc
(like CollectionFetchJob emitting result() twice - once because of
the CommitTransaction response not addressed to it and second time
once the actual CFJ response arrived) in both clients and resources.