Fix read-after-free in CollectionStatistics
CollectionStatistics lives in a separate thread. Returning the Statistics
structure as a reference to other threads than means that the structure
can be deleted in the CollectionStatistics thread while other threads
are still holding a reference. We now return a copy of the Statistics
struct instead, it's just four ints.

Thanks to Andreas Schneider for pointing out the issue.