Port StatisticsProxyModel to KExtraColumnsProxyModel.
* this fixes sibling() - the unittest confirmed Dan's assessment that
sibling() got broken for StatisticsProxyModel with Qt5.

* this reduces the number of dataChanged signals emitted by the proxy:
instead of dataChanged((row,0) (row,0)) + dataChanged((row,1) (row,3))
the proxy now emits dataChanged((row,0) (row,3)) directly

* I also removed the code emitting dataChanged for parent indexes,
the reasoning was "so that cumulative totals can be updated", but there
are no cumulative totals anywhere in this proxy (other than the tooltip).

* the awful model-index-struct-casting hack is no longer necessary,
the code is much simpler, there are now unittests (both here and for
the base class KExtraColumnsProxyModel in kitemmodels)...

BUG: 355229