Update Akonadi documentation, enable docs generation on api.kde.org
This adds metainfo.yaml with metadata for kapidox running on api.kde.org
so we get documentation generated there. It also switches from .dox to
markdown syntax in documentation, as recommended by kapidox.

The high-level documentation is painfully outdated and needs a lot of love.
I fixed only the most obvious issues I noticed (like mentions of Nepomuk
or Akonadi::Entity class), but more things need to be updated and the
documentation needs to be better structured. We should also have some
more examples.